polkadot-useful type definitions for statement table
