+++ polkadot/statement-table/Cargo.toml
@@ -5,3 +5,4 @@ authors = ["Parity Technologies <admin@parity.io>"]
 
 [dependencies]
 substrate-primitives = { path = "../../substrate/primitives" }
+polkadot-primitives = { path = "../primitives" }
